Extra styling and pen pressures
app.js:
Moved default values into js/Defaults.js

css/controls.css:
Removed container around content and controls menu, as grid styling can be applied to body instead.
Removed some redudant styling.

css/graph.css:
Changed the background of the tooltip to match the styling of the buttons in the controls.

css/heatmap.css:
Accomodated change to heatmap naming system.

index.html:
Removed the type="text/javascript" from the script tags, as they are redundant (the type is already specified by the file tag).
Removed the content wrapper from the body, as this can be accounted for by changing styling and is just "a bit nicer".

js/Defaults.js:
Added to store the default values for the webpage.
We can't do this in a json file as the defaults include function definitions.

js/Graph.js:
Moved the y-axis label closer to the axis.
Removed redundant transition to hide the tooltip.

js/Heatmap.js:
changed the drawing style to use mouse pointer events instead of usual mouse events, which allows for touchpad pressure to be measured and used to change the pen size in a more intuitive manner.
Added a kernelShrink method to support this adjustment.